backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
supabase==2.3.4
google-generativeai==0.3.2
Pillow==10.2.0

